인증체크 완료

+++ src/main/java/site/metacoding/white/web/BoardApiController.java
@@ -47,6 +47,9 @@ public class BoardApiController {
   public ResponseDto<?> saveV2(@RequestBody BoardSaveReqDto boardSaveReqDto) {
     // 세션 정보를 들고옴
     SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+    if (sessionUser == null) {
+      throw new RuntimeException("로그인이 필요합니다.");
+    }
     // insert into board(title, content, user_id) values(?, ?, ?)
     boardSaveReqDto.setSessionUser(sessionUser);
     BoardSaveRespDto boardSaveRespDto = boardService.save(boardSaveReqDto); // 서비스에는 단 하나의 객체만 전달한다.
@@ -64,11 +67,14 @@ public class BoardApiController {
   @GetMapping("/board")
   public ResponseDto<?> findAll() {
     return new ResponseDto<>(1, "성공", boardService.findAll());
-
   }
 
   @PutMapping("/board/{id}")
   public ResponseDto<?> update(@PathVariable Long id, @RequestBody BoardUpdateReqDto boardUpdateReqDto) {
+    SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+    if (sessionUser == null) {
+      throw new RuntimeException("로그인이 필요합니다.");
+    }
     // id 따로 주소로 받으니까 board에는 id 받으면 안됨 / board 말고 필요한 데이터만 받아야됨 dto
     boardUpdateReqDto.setId(id);
     return new ResponseDto<>(1, "성공", boardService.update(boardUpdateReqDto));
@@ -76,6 +82,10 @@ public class BoardApiController {
 
   @DeleteMapping("/board/{id}")
   public ResponseDto<?> deleteById(@PathVariable Long id) {
+    SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+    if (sessionUser == null) {
+      throw new RuntimeException("로그인이 필요합니다.");
+    }
     boardService.deleteById(id);
     return new ResponseDto<>(1, "성공", null);
   }
